+++ backend/gonzales/static/assets/HistoryPage-ql7SDVl1.js
@@ -1,4 +1,4 @@
-import{a as m,j as e}from"./query-CSQiZfg4.js";import{c as M,u as _,T as E,X as S,k as N,e as A,l as L,m as F,n as z,S as I}from"./index-Cc5-1RId.js";import{G as R}from"./GlassCard-D-K6o67A.js";import{D as H}from"./DateRangeFilter-BudEyRAF.js";import{f as D}from"./format-Bm42Q-0N.js";import{A as $,a as K}from"./arrow-up-M06ytjfA.js";import"./vendor-7Rr5RJ8N.js";import"./calendar-BlkOUKUw.js";/**
+import{a as m,j as e}from"./query-CSQiZfg4.js";import{c as M,u as _,T as E,X as S,k as N,e as A,l as L,m as F,n as z,S as I}from"./index-DWOQdPpf.js";import{G as R}from"./GlassCard-D-K6o67A.js";import{D as H}from"./DateRangeFilter-CYDVcve-.js";import{f as D}from"./format-Bm42Q-0N.js";import{A as $,a as K}from"./arrow-up-Cq-iCFry.js";import"./vendor-7Rr5RJ8N.js";import"./calendar-TDXs4aDI.js";/**
  * @license lucide-react v0.563.0 - ISC
  *
  * This source code is licensed under the ISC license.