+++ backend/gonzales/static/assets/arrow-up-Cq-iCFry.js
@@ -1,4 +1,4 @@
-import{c as o}from"./index-Cc5-1RId.js";/**
+import{c as o}from"./index-DWOQdPpf.js";/**
  * @license lucide-react v0.563.0 - ISC
  *
  * This source code is licensed under the ISC license.