+++ backend/gonzales/static/assets/trending-down-DcR3n45P.js
@@ -1,4 +1,4 @@
-import{c as n}from"./index-Cc5-1RId.js";/**
+import{c as n}from"./index-DWOQdPpf.js";/**
  * @license lucide-react v0.563.0 - ISC
  *
  * This source code is licensed under the ISC license.